Third_party uses http_proxy env variables
We create a thirdPartyCache derivation to split the downloading from
the autotool reconfiguration of thrift made by fetch_packages. Since
we want to use http_proxy, we need to have a deterministic output
path. However fetch_packages reconfigures thirft and the produced
paths are really sensible to autotool versions (that come from
nixpkgs). Otherwise, each time autotools are upgraded in nixpkgs, the
output path of third party should be changed.