tools.databaseLoader: improvements
Provides changes to make the VM more useful

 * add missing modules for contrail services
 * use all-in-one module in databaseLoader VM so that all contrail
   services can be used
 * only disco and api are started automatically so it doesn't inpact
   boot time
 * all other contrail services are not started but can be with systemctl
 * rabbit, cassandra now listen on 0.0.0.0 for easy port forwarding
 * allow to override QEMU_NET_OPTS of the run script to forward more
   ports
 * contrail-api-cli, cqlsh are included by default
 * allow to load all contrail tables

Not starting automatically schema and svc allows the user to inspect the
DB before running theses services.

To run svc-monitor you need to start first the analytics-api.

Since all services bind to 0.0.0.0 it is easy to redirect ports with
QEMU_NET_OPTS. This is useful if one want to run some contrail service
outside the VM (schema-transformer for example)