Use builtins.fetchTarball for bootstraping
This introduces the use of builtins.fetchTarball for bootstrapping in
nixpkgs-fetch.nix. The previous code relied on the system provided
<nixpkgs> for bootstrapping.

This also works as a drive-by-fix for the occasionally occurring error
message around a missing `.drv` when trying to evaluate queries. Also
the output now has "pinned-18.09-nixpkgs" which can be useful in error
scenarios/debugging.